Feature: tools: Use formatted output in crm_error
Add an XML output option (default: text).

We're allowing description to be NULL in the format functions, even
though crm_error doesn't currently support displaying the name only.
This will make it easy to add such an option if we want to later. But
more importantly, we'll want to allow displaying name and/or description
when we move this functionality to libpacemaker soon.

Ref T95

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>