Low: libpacemaker: Standardize rule check error processing
And write a message even when unable to check a rule. That way we get a
discrete result in XML for every rule that we requested to check.

This also cleans up eval_rule() further, in that it no longer carries
side-effects. We don't have to pass it the out object.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>