Refactor: libcrmcommon: Add pcmk__xe_foreach_child.
This function can be used to iterate over all children of a given XML
node with a given name.  It can also be used to iterate over all
children of a given XML node if no name is given.  This can be used to
replace all the first_named_child/crm_next_same_xml loops, as well as
the pcmk__xe_first_child/pcmk__xe_next loops.

Fixes T525