Fix: tools: Use correct printf/sscanf specs in notifyServicelogEvent.c
U64T and U64TS aren't defined within C. They're defined in our
configure.ac for some reason, but the explanation for why is quite
opaque, and the person who wrote it is no longer an active Pacemaker
contributor.

The standard PRIu64, etc., print/scan specs are already used elsewhere
in Pacemaker, so it should be fine to use them here.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>