Fix: schemas: crm_resource --validate validation fails
In case of an error, the output of `crm_resource --validate` fails to
validate (if validate-all does not output XML). This is because if a
<choice> contains two elements with the same name in a RelaxNG schema,
only the first occurrence is honored and the rest are ignored. (This
does not seem to be documented clearly; it's a conclusion based on
experimentation.)

The solution is to create just one <element> that contains a <choice>
(instead of a <choice> that contains two <element>s).

Closes RHBZ#2123727

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>