Refactor: scheduler: Split resource-config message by output format.
The text formatter is the only one that wants to output the "Resource
XML:" string, and that's really only there to keep it visually separated
from the rest of the resource's output for the end user.  Thus, it makes
sense that the text message should be its own function separate from any
default.

This also eliminates an incorrect use of pcmk__formatted_printf, which
won't do anything for non-text formats anyway.