Refactor: schemas: New element-rule-rsc_defaults definition
At a compatibility break, we'd like to make the rule schema more
fine-grained depending on the context in which the rule occurs.
* <rsc_expression>: This should be allowed only within rsc_defaults or
  op_defaults.

In the meantime, we have to preserve a lot of garbage for compatibility.
The goal of this commit is to make it easier to improve the schema
granularity in the future.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>