Refactor: libpacemaker: Split crmadmin messages by output format.
It's conceivable that these messages could be useful in the log format
or some other future format, in which case the default versions should
not be using pcmk__formatted_printf.  That function is only for use in
writing text message handlers.

Instead, rearrange these messages to add a default that does whatever
the typical operation is.  If ->is_quiet is set, these will do nothing.
The text version can do its own thing if ->is_quiet is set, and
otherwise call the default version.

Unfortunately, most of these are difficult to test without a cluster.