Refactor: libpacemaker: Don't use pcmk__formatted_printf...
...in pcmk__fence_history.  This function can be called for all sorts of
output formats, not just text, so it should not use
pcmk__formatted_printf.

Instead, add an extra argument to the stonith-event message that controls
whether this extremely minimal output format will be used.  This
argument will always be set if ->is_quiet is true, but only the text
handler will look at it.  This should preserve existing output for all
other output formats.