Test: scheduler: Promotable bundle colocations
Currently, bundle colocations aren't considered properly with respect to
promotion, including in the determination of instance ordering. These
tests set a baseline for the behavior of a promotable bundle in various
colocations with a primitive IPaddr2 resource and with another bundle.

Here we add 12 tests covering all combinations of:
* Resource types:
  * Primary bundle promoted instance, dependent primitive
  * Dependent bundle promoted instance, primary primitive
  * Primary and dependent bundle promoted instances
* colocation or anti-colocation
* mandatory or optional

Currently only bundle-promoted-colocation-1 is correct.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>