Test: cts-cli: Fix another test order dependency
When we set PCMK_trace_functions to a particular value in a test group,
we first save the original value and restore it when we're finished
tracing. However, if PCMK_trace_functions is initially undefined, then
we treat the initial value as an empty string and restore it as such.
This of course defines PCMK_trace_functions.

This matters because Pacemaker's error output differs depending on
whether PCMK_trace_functions is undefined or an empty string.
If PCMK_trace_functions is defined (even if empty), the error prefixes
contain file names and line numbers in addition to the function name.

Here we define it as an empty string at the beginning of the script, if
it's not defined.

We also add to the "big sed" all the functions whose line numbers and
file names are now printed, so that the final output doesn't depend on
file name or line number.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>