Low: controller: ignore CRM_OP_LRM_REFRESH
This was only sent by crm_resource --refresh in versions 1.1.9 and earlier.
Since the local crm_resource is the same version as the controller, and
Pacemaker Remote was introduced in 1.1.9, this means that only remote nodes
running 1.1.9 can possibly send it.

It didn't really do anything useful anyway, so just ignore it.