Test: cts-cli: Use temp file for test_assert stdout
Commit 5c0034b7 switched to using a variable to capture the output. It
turns out when the variable holds a very large string, tihs can cause a
"file name too long" error in the echo command. This error gets hidden
by sed (more details in an upcoming commit).

Here we go back to using a temp file for stdout. Note that this doesn't
revert 5c0034b7; we still need to send stderr to a separate temp file,
cat each of them in sequence, and pass only the stdout file to xmllint.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>