Refactor: libcrmcommon: Add timeout arg to pcmk__send_ipc_request.
Passing 0 for the timeout (like is done in almost all places right now)
just does the default, which is five seconds.

Only in pcmk__attrd_api_update is a different value being passed.
There, if one lf the pcmk__node_attr_sync_* flags is set then we pass a
timeout of -1 which means to wait forever.  This is how we're going to
handle the sync points when updating an attribute.  Other operations may
wish to do this in the future.