Feature: various: remove support for versioned agent parameters
This experimental feature was disabled by default and required manually
editing configure.ac to enable it. The decision has been made to drop it
due to low interest in the feature, high complexity of the code, and limited
benefit. Also, changing parameters incompatibly across resource agent versions
is a risky practice to begin with, and should be discouraged.

Closes T469