Refactor: libcrmcommon: clarify pointer arguments
Mark pointer arguments as const where appropriate, and ensure doxygen blocks
use [in] or [in,out] as appropriate.

This is not comprehensive but focuses on functions that have doxygen blocks
and some functions called by them. It also skips acl.c and xml.c for now since
unrelated work in progress will change them significantly.