Test: cts-cli: Remove trailing space from version sed patterns
Fix two occurrences of an issue in the "big sed". The
' api-version="..." ' and ' version="..." ' patterns require a trailing
space in order to match. This doesn't match if our target is the last
attribute in an XML element -- for example:

    <element name="some_name" version="1.0"/>

The failure to match and replace 'version="X">' initially broke CI tests
for an upcoming commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>