Refactor: scheduler: functionize group member internal constraints
group_internal_constraints() was overlong

A couple of pcmk__order_starts() calls are moved for readability. This
theoretically could have some obscure side effects if the relevant actions have
not already been created, but starts should have been created already, and no
regression tests change, so it is highly unlikely that it matters in this case.