Low: fencer: skip "on" in remapped reboot when agent doesn't support it
Previously, when a topology reboot was remapped to "all off then all on",
and a device's agent didn't support "on", we would correctly skip the device
after checking for capable peers, but we would log a warning. To save time and
avoid confusion, skip the device before checking for capable peers.

Fixes T461