Low: fencer: check "on" support when checking whether device can fence
An initiator calls can_fence_host_with_device() to create a list of devices
that can gence a given target. If the requested action is "on", and a device
doesn't support "on", return false immediately rather than wait until the fence
attempt to determine it can't be done.

This allows for better log messages and slightly faster response.

Fixes T511