Feature: daemons: Add support for local sync points on updates.
In the IPC dispatcher for attrd, add the client to a wait list if its
request specifies a sync point.  When the attribute's value is changed
on the local attrd, alert any clients waiting on a local sync point by
then sending the previously delayed ACK.

Sync points for other requests and the global sync point are not yet
supported.

Fixes T35.