Test: cts-scheduler: update scores for colocation change
Using a single implementation for adding colocated node scores resulted in
minor score changes to two regression tests.

For "X with group with Y" colocations, the score for Y on its assigned node is
higher than it was before due to considering the last group member's
colocations instead of the group's (which are empty) when considering the
chained effect of "X with group".