Test: cts-scheduler: add test for anti-colocated groups with failed dependent
Two groups in a two-node cluster are optionally anti-colocated.
The last member of the dependent group is failed and at its migration
threshold.

As of the current code, the scheduler will stop the failed member, rather than
move its group to the same node as its anti-colocated group. The ideal behavior
is debatable, but this gives us reference behavior to compare against.

Ref T217