Feature: daemons: Respond to received attrd confirmation requests.
On the receiving peer side, if the XML request contains confirm="true",
construct a confirmation message after handling the request completes
and send it back to the originating peer.

On the originating peer side, add a skeleton handler for confirmation
messages.  This does nothing at the moment except log it.