Fix: scheduler: prioritize group colocations properly
When assigning a group, the group's "this with" constraints get added to the
first member's, and the group's "with this" constraints get added to the last
member's.

Previously, this was done with simple list concatenation. However, colocations
need to be sorted by the priority of the other resource involved in the
colocation, so use the new functions for this purpose instead.