Refactor: scheduler: prioritize clone colocations properly
This is analogous to 1506bed7 for clones.

When assigning a clone, the clone's colocations get added to each instance.

Previously, this was done with simple prepending. However, colocations
need to be sorted by the priority of the other resource involved in the
colocation, so use the new functions for this purpose instead.

This has no effect, since the clone defers assignment to its colocation
primaries before ever getting to this point. But it's still logically better
and could theoretically have some obscure side effect somewhere for the better.