Feature: daemons: Handle cluster-wide sync points in attrd.
When an attrd receives an IPC request to update some value, record the
protocol versions of all peer attrds.  Additionally register a function
that will be called when all confirmations are received.

The originating IPC cilent (attrd_updater for instance) will sit there
waiting for an ACK until its timeout is hit.

As each confirmation message comes back to attrd, mark it off the list
of peers we are waiting on.  When no more peers are expected, call the
previously registered function.

For attribute updates, this function just sends an ack back to
attrd_updater.

Fixes T35