API: libcrmcommon: New pcmk_pacemakerd_state_remote enum value
This is not technically a pacemakerd state. However, it can be used to
indicate that pacemaker-remoted is running on a Pacemaker Remote node.
pacemaker-remoted, as the main daemon, is the nearest meaningful analog
of pacemakerd on a Pacemaker Remote node.

Attempts to connect to the pacemakerd API on a Pacemaker Remote node
return EREMOTEIO. Callers may find it convenient to use
pcmk_pacemakerd_state_remote as a pseudo-state in this case.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>