Refactor: libpacemaker: Remove pcmk_status.c:cib_connect()
Replace with a call to cib__signon_query().

There is one small side effect of this change: cib__signon_query()
creates a read-write connection, while pcmk_status.c:cib_connect()
creates a read-only connection. In practice, this will probably never
matter.

It certainly doesn't matter now: in both of pcmk__status()'s two
current callers, the CIB client is freed right after the call.

This also means the cib argument of pcmk__status() is currently
unnecessary. Each current caller creates a CIB client just for this
call, so we could just create a client within pcmk__status(). There may
be use for it in the future though, and besides, it's out of scope for
this commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>