Refactor: tools: Use pcmk__query_node_name() in crm_attribute.c
Replace static functions with an internal API call for getting the local
node name from the controller.

It's unclear whether we really want to keep the
get_node_name_from_local() function. It's only used when we're working
with a CIB file (instead of a live CIB) and an explicit node name was
not passed on the command line. In that case, it may not be appropriate
to guess the target node name in the CIB file based on our own hostname.
For example, if the CIB file is from a different cluster, our node name
is "node1", and there's also a "node1" in the other cluster, it doesn't
make sense to use "node1" automatically as the target node name.

That's something to think about but out of scope for now.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>