API: libpacemaker: New pcmk_query_node_info() function
This new function queries the controller to get the node info
corresponding to a particular node ID. This includes node name, UUID,
membership state, whether the node has quorum, and whether the node is
a Pacemaker Remote node. It stores the retrieved values in output
arguments if given, and it outputs them on request.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>