Feature: tools: crm_mon now shows last_update origin
crm_mon shows which node the update came from, when run against a native
(live) CIB. The last_update origin is not reliably meaningful for CIB
files or remote CIB connections and is thus not shown.

For simplicity, the last_update origin is also not shown if for some
reason we fail to get the node name from the controller when using a
native CIB. If desired, we can show "on unknown node" or similar in the
output in that situation.

The new pcmk__node_name global variable in libcrmcommon may be able to
make the *_our_uname variables redundant (for example, fsa_our_uname,
stonith_our_uname), since everything links against libcrmcommon. We ca
evaluate that soon.

Closes T174

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>