Feature: libpe_status: crm_mon shows "maintenance" for rsc maint meta
Previously, if a resource's "maintenance" meta attribute was set to true
(and the cluster's maintenance-mode property was not enabled), the
crm_mon output for the resource said "(unmanaged)". Now, it says
"(maintenance)" instead.

Note that "maintenance" necessarily implies "unmanaged". So as was
already the case with "blocked", if we show "maintenance" then we don't
show "unmanaged".

Note also that this improved flag appears whether the resource is in
maintenance mode because of cluster-wide maintenance mode, because
the "maintenance" resource meta attribute, or both. If the cluster-
wide maintenance-mode property is set, it will be indicated elsewhere
in the crm_mon output, and there will be no separate indication of the
meta attribute until the cluster is taken out of maintenance mode.

Finally, the "(maintenance)" flag does not appear if the resource's node
is in maintenance mode, unless the cluster-wide or resource-level
maintenance attribute is also set. This will be resolved in a future
commit.

Ref T246

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>