Fix: libpe_status: Node maintenance mode sets resource maintenance flag
This seems to have been only a display issue and not a behavioral issue.
Previously, when a resource's node was placed into maintenance mode, the
resource's managed flag was cleared, but its maintenance flag was not
set. As a result, crm_mon showed "(unmanaged)" rather than the more
specific "(maintenance)" for the affected resource.

Node maintenance handling was added in commit 1530e2bf, while the
pe_rsc_maintenance flag was added later in 9a8cb865. The author of that
commit probably just forgot to add a line setting the pe_rsc_maintenance
flag in this situation.

Ref T246

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>