Tests: Fix running pcmk__procfs_pid2path_test on i686.
The expect_value/check_expected_ptr combo from cmocka fails on i686, but
only if the pointer in question is a statically declared buffer.  In
this case, somewhere in the giant pile of casting that occurs in cmocka,
one of the variable or expected value ends up a 32-bit quantity while
the other ends up a 64-bit quantity.  The comparison then fails.

Changing these variables into dynamically allocated buffers makes
everything work out fine.  This is necessary to get builds working
again.