Test: cts-cli: Add tests for maintenance mode crm_mon output
Here we add an XML test for cluster-wide maintenance mode, as well as
text and XML tests for node maintenance and resource maintenance. These
are for comparison after upcoming commits.

The crm-mon-rsc-maint.xml file has the maintenance meta attribute
enabled for all but one resource (the fence resource). This is easier
than fiddling with cibadmin commands in cts-cli.

Ref T246

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>