Low: scheduler: add "group with" colocations to later members independent of assignment
A group's "this with" colocations are added to its later members' colocations when
an earlier member is unmanaged. Previously, this was done only once the group
started assignment, to match the original behavior.

Now, add them regardless of assignment, for consistency.

No regression tests are affected, and it is at least highly unlikely and maybe
impossible to reach this code before assignment.