Test: cts-scheduler: update regression test for colocation change
Now that groups keep their "this with" colocations even after assignment,
one regression test had an inconsequential change.

The promoted-partially-demoted-group test now has one synapse (for promoting
vip-164-fw) whose priority was upgraded to infinity because the instance's
priority is now infinity due to a mandatory colocation of the promoted role of
the resource's group (ms-cdev-pool-0-iscsi-vips-fw) with another resource
(cdev-pool-0-iscsi-vips), which is now properly considered when sorting
instances for promotion.