Low: controller: Node exits fatally in response to join nack
The client already listens for a nack in do_cl_join_finalize_respond(),
but the DC never sends one. Rather, the DC only sends a finalization
message if the node is integrated (acked).

The only change needed at the client end is to exit fatally with no
respawn. Currently, if the client were to receive a nack, it would
respawn and enter a rapid loop (observed after modifying the DC to send
a nack).

The DC needs to send a nack, which means it needs to finalize if there
are **either** integrated nodes **or** nacked nodes. There may be a
better place to put the nack message than the finalize_join_for()
function -- or perhaps the function should be renamed -- but this seems
to work, and there's already a code block there related to nacks.

Closes T453

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>